Optimize GQL GetInitiateVoucherPageResponse.kt

+++ features/merchant/seller_mvc_creation/src/main/java/com/tokopedia/mvc/domain/entity/VoucherCreationMetadata.kt
@@ -5,7 +5,7 @@ data class VoucherCreationMetadata(
     val isEligible: Int,
     val maxProduct: Int,
     val prefixVoucherCode: String,
-    val shopId: Int,
+    val shopId: Long,
     val token: String,
-    val userId: Int
+    val userId: Long
 )
